# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : { },
    "description" : "API for Acontext.",
    "title" : "Acontext API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/api/v1"
  } ],
  "paths" : {
    "/agent_skills" : {
      "get" : {
        "description" : "List all agent skills under a project",
        "parameters" : [ {
          "description" : "User identifier to filter skills",
          "in" : "query",
          "name" : "user",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Limit of agent skills to return, default 20. Max 200.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "Cursor for pagination",
          "in" : "query",
          "name" : "cursor",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Order by created_at descending if true",
          "in" : "query",
          "name" : "time_desc",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_agent_skills_get_200_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "List agent skills",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# List all skills with pagination\nresult = client.skills.list_catalog(limit=50)\nfor skill in result.items:\n    print(f\"{skill.name}: {skill.description}\")\n\n# Paginate through all skills\nif result.has_more:\n    next_page = client.skills.list_catalog(cursor=result.next_cursor)\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// List all skills with pagination\nconst result = await client.skills.list_catalog({ limit: 50 });\nresult.items.forEach(skill => {\n  console.log(`${skill.name}: ${skill.description}`);\n});\n\n// Paginate through all skills\nif (result.has_more) {\n  const nextPage = await client.skills.list_catalog({ cursor: result.next_cursor });\n}\n"
        } ]
      },
      "post" : {
        "description" : "Upload a zip file containing agent skill. The zip file must contain a SKILL.md file (case-insensitive) with YAML format containing 'name' and 'description' fields. The name and description will be extracted from SKILL.md. Files are stored as Disk Artifacts. Optionally associate with a user identifier.",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/_agent_skills_post_request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_agent_skills_post_201_response"
                }
              }
            },
            "description" : "Returns agent skill with name and description extracted from SKILL.md"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Create agent skill",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\nfrom acontext.uploads import FileUpload\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Upload a skill from a zip file\nwith open('my_skill.zip', 'rb') as f:\n    skill = client.skills.create(\n        file=FileUpload(filename='my_skill.zip', content=f.read(), content_type='application/zip'),\n        user='alice@example.com',\n        meta={'version': '1.0'}\n    )\nprint(f\"Created skill: {skill.name} (ID: {skill.id})\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\nimport fs from 'fs';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Upload a skill from a zip file\nconst fileBuffer = fs.readFileSync('my_skill.zip');\nconst skill = await client.skills.create({\n  file: ['my_skill.zip', fileBuffer, 'application/zip'],\n  user: 'alice@example.com',\n  meta: { version: '1.0' }\n});\nconsole.log(`Created skill: ${skill.name} (ID: ${skill.id})`);\n"
        } ]
      }
    },
    "/agent_skills/{id}" : {
      "delete" : {
        "description" : "Delete agent skill and all associated files",
        "parameters" : [ {
          "description" : "Agent skill UUID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Delete agent skill",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Delete a skill by ID\nclient.skills.delete('skill-uuid-here')\nprint('Skill deleted successfully')\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Delete a skill by ID\nawait client.skills.delete('skill-uuid-here');\nconsole.log('Skill deleted successfully');\n"
        } ]
      },
      "get" : {
        "description" : "Get agent skill by its UUID",
        "parameters" : [ {
          "description" : "Agent skill UUID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_agent_skills_post_201_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Get agent skill by ID",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Get a skill by ID\nskill = client.skills.get('skill-uuid-here')\nprint(f\"Skill: {skill.name}\")\nprint(f\"Description: {skill.description}\")\nprint(f\"Files: {len(skill.file_index)} file(s)\")\nfor f in skill.file_index:\n    print(f\"  - {f.path} ({f.mime})\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Get a skill by ID\nconst skill = await client.skills.get('skill-uuid-here');\nconsole.log(`Skill: ${skill.name}`);\nconsole.log(`Description: ${skill.description}`);\nconsole.log(`Files: ${skill.file_index.length} file(s)`);\nskill.file_index.forEach(f => console.log(`  - ${f.path} (${f.mime})`));\n"
        } ]
      }
    },
    "/agent_skills/{id}/download_to_sandbox" : {
      "post" : {
        "description" : "Download all files from an agent skill to a sandbox environment. Files are placed at /skills/{skill_name}/.",
        "parameters" : [ {
          "description" : "Agent skill UUID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/handler.DownloadSkillToSandboxReq"
              }
            }
          },
          "description" : "Download to sandbox request",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_agent_skills__id__download_to_sandbox_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "Skill directory already exists in sandbox"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Download skill to sandbox",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Download skill to sandbox\nresult = client.skills.download_to_sandbox(\n    skill_id='skill-uuid',\n    sandbox_id='sandbox-uuid'\n)\nprint(f\"Success: {result.success}\")\nprint(f\"Skill installed at: {result.dir_path}\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Download skill to sandbox\nconst result = await client.skills.downloadToSandbox('skill-uuid', {\n  sandboxId: 'sandbox-uuid'\n});\nconsole.log(`Success: ${result.success}`);\nconsole.log(`Skill installed at: ${result.dir_path}`);\n"
        } ],
        "x-codegen-request-body-name" : "request"
      }
    },
    "/agent_skills/{id}/file" : {
      "get" : {
        "description" : "Get file content or download URL from agent skill. If the file is text-based (parseable), returns parsed content. Otherwise, returns a presigned download URL.",
        "parameters" : [ {
          "description" : "Agent skill UUID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "File path within the skill (e.g., 'scripts/extract_text.json')",
          "in" : "query",
          "name" : "file_path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "URL expiration in seconds for presigned URL (default 900)",
          "in" : "query",
          "name" : "expire",
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_agent_skills__id__file_get_200_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Get file from agent skill",
        "tags" : [ "agent_skills" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Get a file from a skill (text files return content, binary files return URL)\nfile_resp = client.skills.get_file(\n    skill_id='skill-uuid-here',\n    file_path='scripts/main.py',\n    expire=1800  # URL expires in 30 minutes\n)\n\nprint(f\"File: {file_resp.path} ({file_resp.mime})\")\nif file_resp.content:\n    print(f\"Content: {file_resp.content.raw}\")\nif file_resp.url:\n    print(f\"Download URL: {file_resp.url}\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Get a file from a skill (text files return content, binary files return URL)\nconst fileResp = await client.skills.getFile({\n  skillId: 'skill-uuid-here',\n  filePath: 'scripts/main.py',\n  expire: 1800  // URL expires in 30 minutes\n});\n\nconsole.log(`File: ${fileResp.path} (${fileResp.mime})`);\nif (fileResp.content) {\n  console.log(`Content: ${fileResp.content.raw}`);\n}\nif (fileResp.url) {\n  console.log(`Download URL: ${fileResp.url}`);\n}\n"
        } ]
      }
    },
    "/disk" : {
      "get" : {
        "description" : "List all disks under a project",
        "parameters" : [ {
          "description" : "User identifier to filter disks",
          "in" : "query",
          "name" : "user",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Limit of disks to return, default 20. Max 200.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "Cursor for pagination. Use the cursor from the previous response to get the next page.",
          "in" : "query",
          "name" : "cursor",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Order by created_at descending if true, ascending if false (default false)",
          "in" : "query",
          "name" : "time_desc",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_disk_get_200_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "List disks",
        "tags" : [ "disk" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# List disks\ndisks = client.disks.list(limit=10, time_desc=True)\nfor disk in disks.items:\n    print(f\"Disk: {disk.id}\")\n\n# List disks for a specific user\ndisks = client.disks.list(user='alice@acontext.io', limit=10)\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// List disks\nconst disks = await client.disks.list({ limit: 10, timeDesc: true });\nfor (const disk of disks.items) {\n  console.log(`Disk: ${disk.id}`);\n}\n\n// List disks for a specific user\nconst userDisks = await client.disks.list({ user: 'alice@acontext.io', limit: 10 });\n"
        } ]
      },
      "post" : {
        "description" : "Create a disk group under a project. Optionally associate with a user identifier.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/handler.CreateDiskReq"
              }
            }
          },
          "description" : "CreateDisk payload",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_disk_post_201_response"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Create disk",
        "tags" : [ "disk" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Create a disk\ndisk = client.disks.create()\nprint(f\"Created disk: {disk.id}\")\n\n# Create a disk for a specific user\ndisk = client.disks.create(user='alice@acontext.io')\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Create a disk\nconst disk = await client.disks.create();\nconsole.log(`Created disk: ${disk.id}`);\n\n// Create a disk for a specific user\nconst userDisk = await client.disks.create({ user: 'alice@acontext.io' });\n"
        } ],
        "x-codegen-request-body-name" : "payload"
      }
    },
    "/disk/{disk_id}" : {
      "delete" : {
        "description" : "Delete a disk by its UUID",
        "parameters" : [ {
          "description" : "Disk ID",
          "in" : "path",
          "name" : "disk_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Delete disk",
        "tags" : [ "disk" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Delete a disk\nclient.disks.delete(disk_id='disk-uuid')\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Delete a disk\nawait client.disks.delete('disk-uuid');\n"
        } ]
      }
    },
...
    "/learning_spaces" : {
      "get" : {
        "description" : "List learning spaces with optional user, meta filter, and cursor pagination.",
        "parameters" : [ {
          "description" : "Filter by user identifier",
          "in" : "query",
          "name" : "user",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Limit per page, default 20, max 200",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "Cursor for pagination",
          "in" : "query",
          "name" : "cursor",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Order by created_at descending",
          "in" : "query",
          "name" : "time_desc",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "URL-encoded JSON for JSONB containment filter",
          "in" : "query",
          "name" : "filter_by_meta",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_learning_spaces_get_200_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "List learning spaces",
        "tags" : [ "learning_spaces" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# List learning spaces\nresult = client.learning_spaces.list(limit=20, time_desc=True)\nfor space in result.items:\n    print(f\"{space.id}: {space.meta}\")\n\n# List with user filter\nresult = client.learning_spaces.list(user='alice@acontext.io')\n\n# List with meta filter\nresult = client.learning_spaces.list(\n    filter_by_meta={'version': '1.0'}\n)\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// List learning spaces\nconst result = await client.learningSpaces.list({ limit: 20, timeDesc: true });\nfor (const space of result.items) {\n  console.log(`${space.id}: ${JSON.stringify(space.meta)}`);\n}\n\n// List with user filter\nconst userSpaces = await client.learningSpaces.list({ user: 'alice@acontext.io' });\n\n// List with meta filter\nconst filtered = await client.learningSpaces.list({\n  filterByMeta: { version: '1.0' }\n});\n"
        } ]
      },
      "post" : {
        "description" : "Create a new learning space. Optionally associate with a user.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/handler.CreateLearningSpaceReq"
              }
            }
          },
          "description" : "Create learning space request",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_learning_spaces_post_201_response"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Create learning space",
        "tags" : [ "learning_spaces" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Create a learning space\nspace = client.learning_spaces.create(\n    user='alice@acontext.io',\n    meta={'version': '1.0'}\n)\nprint(f\"Created space: {space.id}\")\n\n# Create a learning space without a user\nspace = client.learning_spaces.create()\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Create a learning space\nconst space = await client.learningSpaces.create({\n  user: 'alice@acontext.io',\n  meta: { version: '1.0' }\n});\nconsole.log(`Created space: ${space.id}`);\n\n// Create a learning space without a user\nconst space2 = await client.learningSpaces.create();\n"
        } ],
        "x-codegen-request-body-name" : "request"
      }
    },
...
    "/learning_spaces/{id}/learn" : {
      "post" : {
        "description" : "Create an async learning record from a session. Initially stays in pending status.",
        "parameters" : [ {
          "description" : "Learning space UUID",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/handler.LearnReq"
              }
            }
          },
          "description" : "Learn request",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_learning_spaces__id__learn_post_201_response"
                }
              }
            },
            "description" : "Created"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Learn from session",
        "tags" : [ "learning_spaces" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Learn from a session\nrecord = client.learning_spaces.learn(\n    'space-uuid',\n    session_id='session-uuid'\n)\nprint(f\"Learning record: {record.id}, status: {record.status}\")\n\n# Wait for learning to complete\nresult = client.learning_spaces.wait_for_learning(\n    'space-uuid',\n    session_id='session-uuid',\n    timeout=120.0\n)\nprint(f\"Final status: {result.status}\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Learn from a session\nconst record = await client.learningSpaces.learn({\n  spaceId: 'space-uuid',\n  sessionId: 'session-uuid'\n});\nconsole.log(`Learning record: ${record.id}, status: ${record.status}`);\n\n// Wait for learning to complete\nconst result = await client.learningSpaces.waitForLearning({\n  spaceId: 'space-uuid',\n  sessionId: 'session-uuid',\n  timeout: 120\n});\nconsole.log(`Final status: ${result.status}`);\n"
        } ],
        "x-codegen-request-body-name" : "request"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "handler.DownloadSkillToSandboxReq" : {
        "properties" : {
          "sandbox_id" : {
            "description" : "Target sandbox ID",
            "type" : "string"
          }
        },
        "required" : [ "sandbox_id" ],
        "type" : "object"
      },
...
      "serializer.Response" : {
        "properties" : {
          "code" : {
            "type" : "integer"
          },
          "error" : {
            "type" : "string"
          },
          "msg" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "_agent_skills_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/service.ListAgentSkillsOutput"
            }
          },
          "type" : "object"
        } ]
      },
      "_agent_skills_post_request" : {
        "properties" : {
          "file" : {
            "description" : "ZIP file containing agent skill. Must contain SKILL.md (case-insensitive) with YAML format: name and description fields.",
            "format" : "binary",
            "type" : "string"
          },
          "user" : {
            "description" : "User identifier to associate with the skill",
            "type" : "string"
          },
          "meta" : {
            "description" : "Additional metadata (JSON string)",
            "type" : "string"
          }
        },
        "required" : [ "file" ],
        "type" : "object"
      },
      "_agent_skills_post_201_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/model.AgentSkills"
            }
          },
          "type" : "object"
        } ]
      },
      "_agent_skills__id__download_to_sandbox_post_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/handler.DownloadSkillToSandboxResp"
            }
          },
          "type" : "object"
        } ]
      },
      "_agent_skills__id__file_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/service.GetFileOutput"
            }
          },
          "type" : "object"
        } ]
      },
...
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "description" : "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  },
...
}